+++ examples/java-maven/pom.xml
@@ -25,12 +25,12 @@
     <version>0.0.1-SNAPSHOT</version>
 
     <properties>
-        <scala.binary.version>2.13</scala.binary.version>
-        <spark.bigtable.artifact>spark-bigtable-scala2.13</spark.bigtable.artifact>
-        <!--
+<!--        <scala.binary.version>2.13</scala.binary.version>-->
+<!--        <spark.bigtable.artifact>spark-bigtable-scala2.13</spark.bigtable.artifact>-->
+
         <spark.bigtable.artifact>spark-bigtable-scala2.12</spark.bigtable.artifact>
         <scala.binary.version>2.12</scala.binary.version>
-        -->
+
         <spark.version>3.5.1</spark.version>
         <reload4j.version>1.7.36</reload4j.version>
         <spark.bigtable.version>0.4.0-SNAPSHOT</spark.bigtable.version>
