+++ examples/scala-sbt/src/main/scala/spark/bigtable/example/customauth/CustomAccessTokenProvider.scala
@@ -1,7 +1,7 @@
-package spark.bigtable.example.customauth
+package spark.bigtable.example.auth
 
 import com.google.auth.oauth2.GoogleCredentials
-import com.google.cloud.spark.bigtable.customauth.{AccessToken, AccessTokenProvider}
+import com.google.cloud.spark.bigtable.auth.{AccessToken, AccessTokenProvider}
 
 import java.io.IOException
 import java.time.Instant
