+++ spark-bigtable-core/src/main/scala/com/google/cloud/spark/bigtable/auth/AccessToken.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.spark.bigtable.customauth
+package com.google.cloud.spark.bigtable.auth
 
 import com.google.auth.oauth2
