+++ spark-bigtable-core/src/main/scala/com/google/cloud/spark/bigtable/auth/AccessTokenProviderCredentials.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2025 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.cloud.spark.bigtable.auth
+
+import com.google.auth.oauth2.{AccessToken, GoogleCredentials}
+import com.google.common.annotations.VisibleForTesting
+
+import java.io.IOException
+
+class AccessTokenProviderCredentials(private val accessTokenProvider: AccessTokenProvider)
+    extends GoogleCredentials {
+  @throws[IOException]
+  override def refreshAccessToken: AccessToken = {
+    accessTokenProvider.refresh()
+    accessTokenProvider.getAccessToken()
+  }
+}
